[commons] return code checking complete. break. [mission5] created roles for lambdas and groups for honeypot. Stuck at implementation.
Former-commit-id: e450510910c2b864b33939a188e6dd30d587ad32

+++ mission2/mission2-destroy.sh
@@ -24,10 +24,22 @@ fi
 echo "[AWS-Secgame] Destroying terraform resources"
 cd resources/terraform
 terraform destroy -auto-approve -var="profile=$SECGAME_USER_PROFILE" -var="id=$SECGAME_USER_ID" -var="ip=$USER_IP"
+#check terraform destroy's return code, act depending on it. 0 is for a flawless execution, 1 means an error has arisen
+if [[ $? != 0 ]]
+then
+	echo "[AWS-Secgame] Non-zero return code on terraform destroy. There might be a problem. Consider destroying by hand (move to /trash/mission2-$SECGAME_USER_ID/resources/terraform and use terraform destroy, or destroy your resources by hand on the console."
+	exit 2
+fi
+
 
 #destroy key pair
 echo "[AWS-Secgame] Deleting key pair."
 aws --profile $SECGAME_USER_PROFILE ec2 delete-key-pair --key-name AWS-secgame-mission2-keypair-Evilcorp-Evilkeypair-$SECGAME_USER_ID
+if [[ $? != 0 ]]
+then
+	echo "[AWS-Secgame] Non-zero return code on keypair destruction. Use aws --profile $USER_SECGAME_PROFILE ec2 describe-key-pairs and delete-key-pair to manually delete the key pair if needed."
+	exit 2
+fi
 
 
 echo "[AWS-Secgame] Mission 2 destroy complete"
